Moved parser and renderer tests into subdirectories
