+++ test/parsers/html2pdf.test.js
@@ -1,0 +1,32 @@
+"use strict";
+
+const fs = require('fs-extra');
+const assert = require('chai').assert;
+
+const html2pdf = require('../../app/parsers/html2pdf');
+const config = require('../config/config');
+
+const testDir = config.testDir;
+
+describe("Html2pdf", function () {
+    const html = "<h1>Test</h1>";
+    const filename = "testFile.pdf";
+
+    it("Basic test", function(done) {
+        this.timeout(5000);
+        fs.readFile(testDir + "/" + filename, (err) => {
+            assert.ok(err);
+            html2pdf(html, testDir + "/" + "testFile", (err, res) => {
+                assert.notOk(err);
+                assert.ok(res);
+                assert.strictEqual(res.filename, process.cwd() + "/" + testDir + "/" + filename);
+                fs.stat(res.filename, (err, res) => {
+                    assert.notOk(err);
+                    assert.ok(res);
+                    assert.ok(res.isFile());
+                    done();
+                });
+            });
+        });
+    });
+});
